==> Code/14_VIS_basic-visuals.Rmd <==
---
title: "Basic Visuals"
author: "Erin Johnston"
date: "12/9/2020"
output: html_document
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE,
                      message = F,
                      warning = F)

```

```{r}

library(tidyverse)
library(here)
library(ggridges)
library(plotrix)

```

```{r}

full_cp_ml <- read_csv(here("Data/output", "2020-11-18_master_II.csv"))

bgovl_full <- full_cp_ml %>%
  filter(species %in% c("BLU", "GPR", "OLV", "VER", "LCD"), area != "FN", year %in% c(2013:2019))

bgovl_vis <- bgovl_full %>%
  group_by(species, drift, area, year, gridcell, site) %>%
  summarise(drift_cpue = sum(cpue)) %>%
  group_by(species, area, year, gridcell, site) %>%
  summarise(cell_cpue = mean(drift_cpue)) %>%
  mutate(year = as.factor(year))


## updating with 2020 data included

full_cp_ml <- read_csv(here("Data", "2021-06-28_full_master_2020.csv"))

bgovl_full <- full_cp_ml %>%
  filter(species %in% c("BLU", "GPR", "OLV", "VER", "LCD"))

bgovl_vis <- bgovl_full %>%
  group_by(species, drift, area, month, day, year, gridcell, site) %>%
  summarise(drift_cpue = sum(cpue)) %>%
  group_by(species, area, month, day, year, gridcell, site) %>%
  summarise(date_cpue = mean(drift_cpue)) %>%
  group_by(species, area, year, gridcell, site) %>%
  summarise(cell_cpue = mean(date_cpue)) %>%
  mutate(year = as.factor(year))


```

```{r}

ggplot() +
  geom_boxplot(data = bgovl_vis, aes(x = year, y = cell_cpue, color = species))+
  scale_color_manual(values = c("#0000CC", "#663300" , "#006699", "#999900", "#CC3300"))+
  facet_grid(site~area)+
  theme_bw()+
  theme(panel.grid = element_blank())

```

## blue

```{r}

blue <- bgovl_vis %>%
  filter(species == "BLU")

ggplot() +
  geom_boxplot(data = blue, aes(x = year, y = cell_cpue, color = species))+
  scale_color_manual(values = c( "#0000CC"))+
  facet_grid(site~area)+
  theme_bw()+
  theme(panel.grid = element_blank())


```

```{r}

## Replace "BLU" with all other species of interest
dens_dat <- full_cp_ml %>%
  filter(species == "BLU", num_caught != 0, size < 22) %>%
  mutate(year = as.factor(year)) %>%
  uncount(num_caught)

dens_dat$area <- factor(dens_dat$area, levels = c("AN", "PL", "BL", "PB"))

dens <- ggplot()+
  geom_density_ridges(data = dens_dat, aes(x = size, y = year, fill = year), 
                      show.legend = F)+
  geom_vline(xintercept = 21, col = "red")+
  scale_fill_manual(values = c("#003399", "#0000FF", "#3366CC","#006699", "#336666",
                               "#339999","#33CCCC", "#33FFFF", "#00FFCC", 
                               "#00FF66", "#00FF33", "#009933", "#006633", "#000000" ))+
  labs(title = "Juvenile Blue Rockfish Size Density") +
  facet_grid(site ~ area) +
  theme_bw()+
  theme(panel.grid = element_blank())

dens

ggsave("blu_dens.png", plot = dens, path = here("Images/figure_dump"), dpi = 1000, width = 8, height = 6)

dens_dat %>%
  filter(size < 22, area == "PB", year %in% c(2017, 2020)) %>%
  ggplot()+
  geom_bar(aes(x = size))+
  facet_grid(year ~ site)

dens_dat %>%
  filter(size < 22, area == "AN", year %in% c(2014, 2016, 2017)) %>%
  ggplot()+
  geom_bar(aes(x = size))+
  facet_grid(year ~ site)

dens_dat %>%
  filter(size < 22, area == "BL", year %in% c( 2016)) %>%
  ggplot()+
  geom_bar(aes(x = size))+
  facet_grid(year ~ site)

dens_dat %>%
  filter(size < 22, area == "PL", year %in% c( 2017)) %>%
  ggplot()+
  geom_bar(aes(x = size))+
  facet_grid(year ~ site)
```


```{r}
dens_dat_g <- full_cp_ml %>%
  filter(species == "GPR", num_caught != 0) %>%
  mutate(year = as.factor(year))

ggplot()+
  geom_density_ridges(data = dens_dat_g, aes(x = size, y = year, fill = year),
                      show.legend = F)+
  facet_wrap(.~area) +
  theme_classic()
```


```{r}
dens_dat_o <- full_cp_ml %>%
  filter(species == "OLV", num_caught != 0) %>%
  mutate(year = as.factor(year))

ggplot()+
  geom_density_ridges(data = dens_dat_o, aes(x = size, y = year, fill = year))
```


```{r}
dens_dat_black <- full_cp_ml %>%
  filter(species == "BLA", num_caught != 0) %>%
  mutate(year = as.factor(year))

ggplot()+
  geom_density_ridges(data = dens_dat_black, aes(x = size, y = year, fill = year))
```


```{r}
dens_dat_v <- full_cp_ml %>%
  filter(species == "VER", num_caught != 0) %>%
  mutate(year = as.factor(year))

ggplot()+
  geom_density_ridges(data = dens_dat_v, aes(x = size, y = year, fill = year))
```


```{r}
dens_dat_l <- full_cp_ml %>%
  filter(species == "LCD", num_caught != 0) %>%
  mutate(year = as.factor(year))

ggplot()+
  geom_density_ridges(data = dens_dat_l, aes(x = size, y = year, fill = year))
```


Where are olive rf in AN REF?

```{r}

missingolv <- full_cp_ml %>%
  filter(area == "AN", species == "OLV", year  %in% c( 2007:2012), num_caught > 0, site == "REF")

```


## t test visuals

```{r}

t_test_vis <- dens_dat %>%
  group_by(area, site, year) %>%
  summarise(mean_size = mean(size),
            se_size = std.error(size))

t_test_vis %>%
  filter(area == "BL", year == 2016) %>%
  ggplot() +
  geom_col(aes( x = site, y = mean_size), fill = "#CCCCCC", color = "#000000")+
  geom_errorbar(aes(x = site, y = mean_size, ymin = mean_size -se_size,
                    ymax = mean_size + se_size), width = 0.2, size = 0.6) +
  labs(x = "Piedras Blancas", y = "Mean Size")+
  theme_bw()+
  theme(panel.grid = element_blank())

t_test_vis %>%
  filter(area == "AN", year %in% c(2008,2012, 2016, 2017, 2018)) %>%
  ggplot() +
  geom_col(aes( x = site, y = mean_size), fill = "#CCCCCC", color = "#000000")+
  geom_errorbar(aes(x = site, y = mean_size, ymin = mean_size -se_size,
                    ymax = mean_size + se_size), width = 0.2, size = 0.6) +
  facet_wrap(.~year)+
  ##look up unicode for special characters
  labs(x = "A\U00F1o Nuevo", y = "Mean Size")+
  theme_bw()+
  theme(panel.grid = element_blank())
  
t_test_vis %>%
  filter(area == "PL", year == 2017) %>%
  ggplot() +
  geom_col(aes( x = site, y = mean_size), fill = "#CCCCCC", color = "#000000")+
  geom_errorbar(aes(x = site, y = mean_size, ymin = mean_size -se_size,
                    ymax = mean_size + se_size), width = 0.2, size = 0.6) +
  labs(x = "Point Lobos", y = "Mean Size")+
  theme_bw()+
  theme(panel.grid = element_blank())
  

t_test_vis %>%
  filter(area == "PB", year %in% c(2017, 2020)) %>%
  ggplot() +
  geom_col(aes( x = site, y = mean_size), fill = "#CCCCCC", color = "#000000")+
  geom_errorbar(aes(x = site, y = mean_size, ymin = mean_size -se_size,
                    ymax = mean_size + se_size), width = 0.2, size = 0.6) +
  facet_wrap(.~year)+
  ##look up unicode for special characters
  labs(x = "Point Buchon", y = "Mean Size")+
  theme_bw()+
  theme(panel.grid = element_blank())



```
