==> Code/31_blue_cross_corr_2020.Rmd <==
---
title: "Blue Rockfish Cross Correlation"
author: "Erin Johnston"
date: "6/23/2021"
output: 
  prettydoc::html_pretty:
    theme: cayman
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = F)

```

[back](https://erinmarjo.github.io/CCFRP/27_thesis_progress_report.html)

```{r libraries, warning=F, message=F}

library(tidyverse)
library(here)
library(knitr) ## for kable()
library(kableExtra) ## for save_kable()

```


```{r}

master <- read_csv(here("Data", "2021-06-28_full_master_2020.csv"))

blue_full <- master %>%
  filter(species == "BLU") 

missing_bl_blue <- data.frame(area = c("BL", "BL", "BL", "BL"),
                              site = c("REF", "MPA", "REF", "MPA"),
                              year = c(2007, 2007, 2015, 2015), 
                              cpue_site = c(NA, NA, NA, NA))

juv <- blue_full %>%
  filter(size <= 21) %>%
  select(-size) %>%
  group_by(drift, trip, area, site, month, day, year, gridcell) %>%
  summarise(cpue_sum = sum(cpue))%>%
  group_by(trip, area, site, month, day, year, gridcell) %>%
  summarise(cpue_date = mean(cpue_sum))%>%
  group_by( area, site, year, gridcell) %>%
  summarise(cpue_cell = mean(cpue_date))%>%
  group_by( area, site, year) %>%
  summarise(cpue_site = mean(cpue_cell)) %>%
  bind_rows(missing_bl_blue)


adult <- blue_full %>%
  filter(size >= 32) %>%
  select(-size)%>%
  group_by(drift, trip, area, site, month, day, year, gridcell) %>%
  summarise(cpue_sum = sum(cpue))%>%
  group_by(trip, area, site, month, year, gridcell) %>%
  summarise(cpue_date = mean(cpue_sum)) %>%
  group_by( area, site, year, gridcell) %>%
  summarise(cpue_cell = mean(cpue_date)) %>%
  group_by( area, site, year) %>%
  summarise(cpue_site = mean(cpue_cell))%>%
  bind_rows(missing_bl_blue)


## interesting thought to see where the fish in between adult and juvenile end up

# teen <- blue_full %>%
#   filter(size > 21 & size < 32) %>%
#   select(-size)%>%
#   group_by(drift, trip, area, site, month, day, year, gridcell) %>%
#   summarise(cpue_sum = sum(cpue))%>%
#   group_by(trip, area, site, month, year, gridcell) %>%
#   summarise(cpue_date = mean(cpue_sum)) %>%
#   group_by( area, site, year, gridcell) %>%
#   summarise(cpue_cell = mean(cpue_date)) %>%
#   group_by( area, site, year) %>%
#   summarise(cpue_site = mean(cpue_cell))%>%
#   bind_rows(missing_bl_blue)

```


```{r data with na}

MOCI <- read_csv(here("Data", "Central_California_MOCI_2020.csv"))

juv_moci <- left_join(juv, MOCI, by = "year") 
adult_moci <- left_join(adult, MOCI, by = "year")

# teen_moci <- left_join(teen, MOCI, by = "year")

## just want to see something
# juv_moci <- juv_moci %>%
#   filter(year %in% c(2007:2019))

```


```{r write csvs, echo = F, eval = F}

#write_csv()

# write_csv(juv_moci, here("Data", "2021-07-04_juv_blue_moci_site_cpue.csv"))
# write_csv(adult_moci, here("Data", "2021-07-04_adult_blue_moci_site_cpue.csv"))

```

```{r, eval = F, echo = F}

## checking to make sure that adding NA's as a different dataframe was the same as manually adding NA's for Piedras Blancas in both 2015 and 2007 for the REF and MPA.
## It is. So there, I can do so for the other species as well.

# juv2 <- blue_full %>%
#   filter(size <= 21) %>%
#   select(-size) %>%
#   group_by(drift, trip, area, site, month, day, year, gridcell) %>%
#   summarise(cpue_sum = sum(cpue))%>%
#   group_by(trip, area, site, month, day, year, gridcell) %>%
#   summarise(cpue_date = mean(cpue_sum))%>%
#   group_by( area, site, year, gridcell) %>%
#   summarise(cpue_cell = mean(cpue_date))%>%
#   group_by( area, site, year) %>%
#   summarise(cpue_site = mean(cpue_cell))
# 
# write_csv(juv2, here("Data", "2021-07-04_juv_blue_site_cpue.csv"))
# juv_na <- read_csv(here("Data", "2021-07-04_juv_blue_site_cpue_NA.csv"))
# 
# juv_moci2 <- left_join(juv_na, MOCI, by = "year") 
# 
# anti_juv <- juv_moci %>%
#   unite(col = "joined", c("area", "site", "year", "season"), sep = "_")
# 
# anti_juv2 <- juv_moci2 %>%
#   unite(col = "joined", c("area", "site", "year", "season"), sep = "_")
# 
# different <- anti_join(anti_juv, anti_juv2, by = "joined")

```


### Cross Correlation Function

```{r custom function}
erin_ccf <- function(full_x, full_y, lag_x_rm, lag_y_rm) {
  mean_x <- mean(full_x, na.rm = T)
  mean_y <- mean(full_y, na.rm = T)
  corval <- (sum((lag_x_rm - mean_x)*
                   (lag_y_rm - mean_y)))/
    ((sqrt(sum((full_x - mean_x)^2, na.rm = T)))*
       (sqrt(sum((full_y - mean_y)^2, na.rm = T))))
  return(corval)
  
}
```


```{r loop set up}

juv_corr <- data.frame(area = character(), site = character(), season = character(), lag = character(), correlation = character())

adult_corr <- data.frame(area = character(), site = character(), season = character(), lag = character(), correlation = character())

# teen_corr <- data.frame(area = character(), site = character(), season = character(), lag = character(), correlation = character())

area_list <- c("PB", "BL", "AN", "PL")
site_list <- c("MPA", "REF")
season_list <- c("JFM", "AMJ", "JAS", "OND")

```

## Juvenile Blue Rockfish

```{r juv loop}

for(j in area_list){
  for(k in site_list){
    for(l in season_list){
      juvenile <- juv_moci %>%
        filter(area == j, site == k, season == l)
      for(i in 0:8){
        lag_df <- juvenile %>%
          mutate(central_ca = dplyr::lag(central_ca, n = i)) %>% drop_na()
        cor <- erin_ccf(juvenile$central_ca, juvenile$cpue_site,
                        lag_df$central_ca, lag_df$cpue_site)
        for(m in 1:length(cor)){
          juv_corr <- bind_rows(juv_corr, c(area = j, site = k, season = l, 
                                            lag = -i, correlation = cor[m]))
        }
      }
    }
  }
} 


juv_vis <- juv_corr %>%
  mutate(correlation = as.numeric(correlation),
         lag = as.numeric(lag),
         Season = case_when(
    season == "JFM" ~ "Winter",
    season == "AMJ" ~ "Spring",
    season == "JAS" ~ "Summer",
    season == "OND" ~ "Fall"
  ), Area = case_when(
    area == "AN" ~ "A\u00f1o Nuevo",
    area == "PL" ~ "Point Lobos",
    area == "BL" ~ "Piedras Blancas",
    area == "PB" ~ "Point Buchon"
  ))
  
juv_vis$season <- factor(juv_vis$season, levels = c("JFM", "AMJ" ,"JAS", "OND"))
juv_vis$Season <- factor(juv_vis$Season, levels = c("Winter", "Spring" ,"Summer", "Fall"))  
juv_vis$Area <- factor(juv_vis$Area, levels = c("A\u00f1o Nuevo", "Point Lobos" ,"Piedras Blancas", "Point Buchon"))  


juv_plot_erin_ccf <- ggplot()+
  scale_color_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  scale_fill_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  geom_col(position = "dodge2", data = juv_vis, aes(x = lag, y = correlation, 
                                                    color = Season, fill = Season))+
  labs(title = "Juvenile Blue Rockfish Lag Correlation")+
  scale_y_continuous(limits = c(-0.6, 1.0), 
                     breaks = c(-0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0))+
  theme_bw()+
  theme(panel.grid = element_blank(),
        plot.title = element_text(hjust = 0.5))+
  facet_grid(site~Area)
  
  
  
juv_simplified <- ggplot()+
  scale_color_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  scale_fill_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  geom_col(position = "dodge2", data = juv_vis, aes(x = lag, y = correlation, 
                                                    color = Season, fill = Season))+
  labs(title = "Juvenile Blue Rockfish Lag Correlation")+
  scale_y_continuous(limits = c(-0.6, 1.0), 
                     breaks = c(-0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0))+
  theme_bw()+
  theme(panel.grid = element_blank(),
        plot.title = element_text(hjust = 0.5))


```

```{r juv simplified}
juv_corr_simp <- data.frame( season = character(), lag = character(), correlation = character())

#for(j in area_list){
#  for(k in site_list){
    for(l in season_list){
      juvenile <- juv_moci %>%
        filter(season == l)
      for(i in 0:8){
        lag_df <- juvenile %>%
          mutate(central_ca = dplyr::lag(central_ca, n = i)) %>% drop_na()
        cor <- erin_ccf(juvenile$central_ca, juvenile$cpue_site,
                        lag_df$central_ca, lag_df$cpue_site)
        for(m in 1:length(cor)){
          juv_corr_simp <- bind_rows(juv_corr_simp, c(area = j, site = k, season = l, 
                                            lag = -i, correlation = cor[m]))
        }
      }
    }
 # }
#}

juv_vis_simp <- juv_corr_simp %>%
  mutate(correlation = as.numeric(correlation),
         lag = as.numeric(lag),
         Season = case_when(
    season == "JFM" ~ "Winter",
    season == "AMJ" ~ "Spring",
    season == "JAS" ~ "Summer",
    season == "OND" ~ "Fall"
  ), Area = case_when(
    area == "AN" ~ "A\u00f1o Nuevo",
    area == "PL" ~ "Point Lobos",
    area == "BL" ~ "Piedras Blancas",
    area == "PB" ~ "Point Buchon"
  ))

juv_vis_simp$Season <- factor(juv_vis_simp$Season, levels = c("Winter", "Spring" ,"Summer", "Fall"))  
juv_vis_simp$Area <- factor(juv_vis_simp$Area, levels = c("A\u00f1o Nuevo", "Point Lobos" ,"Piedras Blancas", "Point Buchon")) 

juv_pooled <- ggplot()+
  scale_color_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  scale_fill_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  geom_col(position = "dodge2", data = juv_vis_simp, aes(x = lag, y = correlation, 
                                                    color = Season, fill = Season))+
  labs(title = "Juvenile Blue Rockfish Lag Correlation")+
  scale_y_continuous(limits = c(-0.6, 1.0), 
                     breaks = c(-0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0))+
  theme_bw()+
  theme(panel.grid = element_blank(),
        plot.title = element_text(hjust = 0.5))

str(juv_corr_simp)


```


### Cross Correlation Figure
#### Juvenile Blue Rockfish vs MOCI

```{r juv result and figure}

juv_vis %>%
  arrange(desc(abs(correlation))) %>%
  group_by(area, site) %>%
  slice(1)


juv_plot_erin_ccf
juv_simplified

```

### Highest Correlation Figure

```{r top juv figure}

top_juv <- juv_vis %>% 
  arrange(desc(correlation)) %>% 
  group_by(area, site) %>% slice(1)



top_juv_meeting <- top_juv %>%
  mutate(Season = case_when(
    season == "JFM" ~ "Winter",
    season == "AMJ" ~ "Spring",
    season == "JAS" ~ "Summer",
    season == "OND" ~ "Fall"
  ), Lag = case_when(
    lag == "-2" ~ "2",
    lag == "-1" ~ "1"
  ), Area = case_when(
    area == "AN" ~ "A\u00f1o Nuevo",
    area == "PL" ~ "Point Lobos",
    area == "BL" ~ "Piedras Blancas",
    area == "PB" ~ "Point Buchon"
  ))

#top_juv_meeting$area <- factor(top_juv_meeting$area, levels = c("AN", "PL", "BL", "PB"))
top_juv_meeting$Area <- factor(top_juv_meeting$Area, levels = c("A\u00f1o Nuevo", "Point Lobos", "Piedras Blancas", "Point Buchon"))
top_juv_meeting$Season <- factor(top_juv_meeting$Season, 
                                 levels = c("Winter", "Spring", "Summer", "Fall"))


top_juv_2020_dat <- top_juv_meeting %>%
  ggplot()+
  geom_point(size = 4, aes(x = Area, y = correlation, color = Season, 
                           fill = Season, shape = Lag))+
  scale_color_manual(values = c("#009999","#666666", "#000000"))+
  scale_fill_manual(values = c("#009999", "#666666", "#000000"))+
  scale_shape_manual(values = c(15, 19))+
  labs(title = "Juvenile Blue Rockfish Top Correlation CPUE-MOCI", 
       y = "Correlation\n", x = "\nArea (N to S)", caption = "")+
  scale_y_continuous(limits = c(0.0, 1.0), 
                     breaks = c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0))+
  facet_wrap(.~site)+
  theme_bw()+
  theme(panel.grid = element_blank(),
        plot.title = element_text(hjust = 0.5),
        plot.caption = element_text(size = 8),
        axis.text.x = element_text(face = "bold",angle = 45, hjust =1))+
  guides(color = guide_legend(override.aes = list(shape = 15, size = 7),
                              order = 1),
         shape = guide_legend(order = 2),
         fill = FALSE)



top_juv_2020_dat


```

```{r hand calculation, echo = F, eval = F}

## checking with ccf

# juv_corr2 <- data.frame(area = character(), site = character(), season = character(), lag = integer(), correlation = character())
# 
# area_list2 <- c("PB", "AN", "PL")
# 
# for(j in area_list2){
#   for(k in site_list){
#     for(l in season_list){
#       juvenile <- juv_moci %>%
#         filter(area == j, site == k, season == l)
#       for(i in 0:8){
#         cor <- ccf(juvenile$central_ca, juvenile$cpue_site)
#         for(m in 1:length(cor)){
#           juv_corr2 <- bind_rows(juv_corr2, c(area = j, site = k, season = l, 
#                                             lag = -i, correlation = cor[m]))
#         }
#       }
#     }
#   }
# }
# 
# by_hand <- juv_moci %>%
#   filter(season == "JAS", area == "AN", site == "MPA")

```

## Significant Correlations: Juvenile

```{r significant correlations juvenile}

juv_sig_corr <- juv_vis %>% 
  arrange(desc(correlation)) %>%
  mutate(pval = 2 * (1 - pnorm(abs(correlation), mean = 0 , sd = 1/sqrt(14)))) %>%
  filter(pval <= 0.05) %>%
  mutate(Season = case_when(
    season == "JFM" ~ "Winter",
    season == "AMJ" ~ "Spring",
    season == "JAS" ~ "Summer",
    season == "OND" ~ "Fall"), 
    Lag = case_when(
    lag == "-3" ~ "3",
    lag == "-2" ~ "2",
    lag == "-1" ~ "1"),
    Area = case_when(
    area == "AN" ~ "A\u00f1o Nuevo",
    area == "PL" ~ "Point Lobos",
    area == "BL" ~ "Piedras Blancas",
    area == "PB" ~ "Point Buchon"))

juv_sig_corr$Area <- factor(juv_sig_corr$Area, levels = c("A\u00f1o Nuevo", "Point Lobos",
                                                          "Piedras Blancas", "Point Buchon"))
juv_sig_corr$Season <- factor(juv_sig_corr$Season, levels = c("Winter", "Spring", 
                                                              "Summer", "Fall"))


juv_sig_corr_plot <- juv_sig_corr %>%
  ggplot()+
  geom_point(size = 4, aes(x = Area, y = correlation, color = Season, 
                           fill = Season, shape = Lag),
             position = position_jitter(width = 0.1, height = 0.1, seed = 69))+ #nice
  scale_color_manual(values = c("#009999","#666666", "#000000"))+
  scale_fill_manual(values = c("#009999", "#666666", "#000000"))+
  scale_shape_manual(values = c(15, 19, 17))+
  labs(title = "Juvenile Blue Rockfish Significant Correlations", 
       y = "Correlation\n", x = "\nArea (N to S)", caption = "")+
  scale_y_continuous(limits = c(0.0, 1.0), 
                     breaks = c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0))+
  facet_wrap(.~site)+
  theme_bw()+
  theme(panel.grid = element_blank(),
        plot.title = element_text(hjust = 0.5),
        plot.caption = element_text(size = 8),
        axis.text.x = element_text(face = "bold",angle = 45, hjust =1))+
  guides(color = guide_legend(override.aes = list(shape = 15, size = 7)))

```

### Tables for pub - juv

```{r juv table for pub}

library(sjPlot)

juv_table <- juv_vis %>%
  select(-Season, -Area) %>%
  mutate(season = as.character(season))

tab_df(juv_table, sort.column = -5, digits = 4, alternate.rows = T,
       file = here("Output", "juv_corr_table.doc"))

top_juv_table <- top_juv_meeting %>%
  select(-Season, -Area, -Lag) %>%
  mutate(season = as.character(season))


tab_df(top_juv_table, sort.column = 2, digits = 4, alternate.rows = T, footnote = "Table X. Highest cross correlation coefficient for juvenile Blue rockfish", show.footnote = T, file = here("Output", "top_juv_corr_table.doc"))
       
# Highest cross correlation \ncoefficients for juvenile Blue rockfish CPUE and MOCI blah blah blah blah blah which shows that blah de blah blah       
#)



```



## Adult Blue Rockfish


```{r adult loop}
for(j in area_list){
  for(k in site_list){
    for(l in season_list){
      adult <- adult_moci %>%
        filter(area == j, site == k, season == l)
      for(i in 0:8){
        lag_df <- adult %>%
          mutate(central_ca = dplyr::lag(central_ca, n = i)) %>% drop_na()
        cor <- erin_ccf(adult$central_ca, adult$cpue_site,
                        lag_df$central_ca, lag_df$cpue_site)
        for(m in 1:length(cor)){
          adult_corr <- bind_rows(adult_corr, c(area = j, site = k, season = l, 
                                            lag = -i, correlation = cor[m]))
        }
      }
    }
  }
} 



adult_vis <- adult_corr %>%
  mutate(correlation = as.numeric(correlation),
         lag = as.numeric(lag),
         Season = case_when(
    season == "JFM" ~ "Winter",
    season == "AMJ" ~ "Spring",
    season == "JAS" ~ "Summer",
    season == "OND" ~ "Fall"
  ), Area = case_when(
    area == "AN" ~ "A\u00f1o Nuevo",
    area == "PL" ~ "Point Lobos",
    area == "BL" ~ "Piedras Blancas",
    area == "PB" ~ "Point Buchon"
  ))
  
adult_vis$season <- factor(adult_vis$season, levels = c("JFM", "AMJ" ,"JAS", "OND"))  
adult_vis$Season <- factor(juv_vis$Season, levels = c("Winter", "Spring" ,"Summer", "Fall"))  
adult_vis$Area <- factor(juv_vis$Area, levels = c("A\u00f1o Nuevo", "Point Lobos" ,"Piedras Blancas", "Point Buchon")) 

adult_plot_erin_ccf <- ggplot()+
  scale_color_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  scale_fill_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  geom_col(position = "dodge2", data = adult_vis, aes(x = lag, y = correlation,
                                                      color = Season, fill = Season))+
  labs(title = "Adult Blue Rockfish Lag Correlation")+
  scale_y_continuous(limits = c(-0.6, 1.0), 
                     breaks = c(-0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0))+
  theme_bw()+
  theme(panel.grid = element_blank(),
        plot.title = element_text(hjust = 0.5))+
  facet_grid(site~Area)

```

### Cross Correlation Figure
#### Adult Blue Rockfish vs MOCI

```{r adult result and figure}

adult_vis %>%
  arrange(desc(correlation)) %>%
  group_by(area, site) %>%
  slice(1)

adult_plot_erin_ccf

```

### Highest Correlation Figure

```{r top adult figure}

top_adult <- adult_vis %>% 
  arrange(desc(correlation)) %>% 
  group_by(area, site) %>% slice(1)


top_adult_meeting <- top_adult %>%
  mutate(Season = case_when(
    season == "JFM" ~ "Winter",
    season == "AMJ" ~ "Spring",
    season == "JAS" ~ "Summer",
    season == "OND" ~ "Fall"
  ), Lag = case_when(
    lag == "0" ~ "0",
    lag == "-1" ~ "1",
    lag == "-2" ~ "2",
    lag == "-3" ~ "3",
    lag == "-4" ~ "4"
  ), Area = case_when(
    area == "AN" ~ "A\u00f1o Nuevo",
    area == "PL" ~ "Point Lobos",
    area == "BL" ~ "Piedras Blancas",
    area == "PB" ~ "Point Buchon"
  )) 

top_adult_meeting$area <- factor(top_adult_meeting$area, levels = c("AN", "PL", "BL", "PB"))
top_adult_meeting$Area <- factor(top_adult_meeting$Area, levels = c("A\u00f1o Nuevo", "Point Lobos", "Piedras Blancas", "Point Buchon"))
top_adult_meeting$Season <- factor(top_adult_meeting$Season, 
                                 levels = c("Winter", "Spring", "Summer", "Fall"))



top_adult_2020_dat <-  top_adult_meeting %>%
  ggplot()+
  geom_point(size = 4, aes(x = Area, y = correlation, 
                           color = Season, fill = Season, shape = Lag))+
  scale_color_manual(values = c("#009999","#666666", "#000000"))+
  scale_fill_manual(values = c( "#009999","#666666", "#000000"))+
  scale_shape_manual(values = c(18, 15, 19, 17))+
  labs(title = "Adult Blue Rockfish Top Correlation CPUE-MOCI", y = "Correlation\n",
       x = "\nArea (N to S)", caption = "")+
  scale_y_continuous(limits = c(0.0, 1.0), 
                     breaks = c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0))+  
  facet_wrap(.~site)+
  theme_bw()+
  theme(panel.grid=element_blank(),
        plot.title = element_text(hjust = 0.5),
        plot.caption = element_text(size = 8),
        axis.text.x = element_text(face = "bold",angle = 45, hjust =1))+
  guides(color = guide_legend(override.aes = list(shape = 15, size = 7)))

top_adult_2020_dat

```


```{r adult table for pub}


library(sjPlot)

top_adult_table <- top_adult_meeting %>%
  select(-Season, -Area, -Lag) %>%
  mutate(season = as.character(season),
         area = as.character(area))


tab_df(top_adult_table, sort.column = 2, digits = 4, alternate.rows = T, footnote = "Table X. Highest cross correlation coefficient for adult Blue rockfish", show.footnote = T, file = here("Output", "top_adult_corr_table.doc"))

```


## Significant Correlations: Adult

```{r significant correlations adult}

adult_sig_corr <- adult_vis %>% 
  arrange(desc(correlation)) %>%
  mutate(pval = 2 * (1 - pnorm(abs(correlation), mean = 0 , sd = 1/sqrt(14)))) %>%
  filter(pval <= 0.05) %>%
  mutate(Season = case_when(
    season == "JFM" ~ "Winter",
    season == "AMJ" ~ "Spring",
    season == "JAS" ~ "Summer",
    season == "OND" ~ "Fall"), 
    Lag = case_when(
    lag == "-5" ~ "5",
    lag == "-4" ~ "4",
    lag == "-3" ~ "3",
    lag == "-2" ~ "2",
    lag == "-1" ~ "1",
    lag == "0" ~ "0"),
    Area = case_when(
    area == "AN" ~ "A\u00f1o Nuevo",
    area == "PL" ~ "Point Lobos",
    area == "BL" ~ "Piedras Blancas",
    area == "PB" ~ "Point Buchon"))

adult_sig_corr$Area <- factor(adult_sig_corr$Area, levels = c("A\u00f1o Nuevo", "Point Lobos",
                                                          "Piedras Blancas", "Point Buchon"))
adult_sig_corr$Season <- factor(adult_sig_corr$Season, levels = c("Winter", "Spring", 
                                                              "Summer", "Fall"))


adult_sig_corr_plot <- adult_sig_corr %>%
  ggplot()+
  geom_point(size = 4, aes(x = Area, y = correlation, color = Season, 
                           fill = Season, shape = Lag),
             position = position_jitter(width = 0.05, height = 0.05, seed = 50))+ #nice
  scale_color_manual(values = c("#999999", "#009999","#666666", "#000000"))+
  scale_fill_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  scale_shape_manual(values = c(18, 15, 19, 17, 4, 25))+
  labs(title = "Adult Blue Rockfish Significant Correlations", 
       y = "Correlation\n", x = "\nArea (N to S)", caption = "")+
  scale_y_continuous(limits = c(0.0, 1.0), 
                     breaks = c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0))+
  facet_wrap(.~site)+
  theme_bw()+
  theme(panel.grid = element_blank(),
        plot.title = element_text(hjust = 0.5),
        plot.caption = element_text(size = 8),
        axis.text.x = element_text(face = "bold",angle = 45, hjust =1))+
  guides(color = guide_legend(override.aes = list(shape = 15, size = 7),
                              order = 1),
         shape = guide_legend(order = 2),
         fill = "none")

```


## Adult CCF Simplified Plots

```{r adult simplified}

adult_blue_no_facet_grid <- ggplot()+
  scale_color_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  scale_fill_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  geom_col(position = "dodge2", data = adult_vis, aes(x = lag, y = correlation,
                                                      color = Season, fill = Season))+
  labs(title = "Adult Blue Rockfish Lag Correlation")+
  scale_y_continuous(limits = c(-0.6, 1.0), 
                     breaks = c(-0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0))+
  theme_bw()+
  theme(panel.grid = element_blank(),
        plot.title = element_text(hjust = 0.5))


```


```{r adult simplified 2}

adult_corr_simp <- data.frame( season = character(), lag = character(), correlation = character())

#for(j in area_list){
#  for(k in site_list){
    for(l in season_list){
      adult <- adult_moci %>%
        filter(season == l)
      for(i in 0:8){
        lag_df <- adult %>%
          mutate(central_ca = dplyr::lag(central_ca, n = i)) %>% drop_na()
        cor <- erin_ccf(adult$central_ca, adult$cpue_site,
                        lag_df$central_ca, lag_df$cpue_site)
        for(m in 1:length(cor)){
          adult_corr_simp <- bind_rows(adult_corr_simp, c(area = j, site = k, season = l, 
                                            lag = -i, correlation = cor[m]))
        }
      }
    }
 # }
#}

adult_vis_simp <- adult_corr_simp %>%
  mutate(correlation = as.numeric(correlation),
         lag = as.numeric(lag),
         Season = case_when(
    season == "JFM" ~ "Winter",
    season == "AMJ" ~ "Spring",
    season == "JAS" ~ "Summer",
    season == "OND" ~ "Fall"
  ), Area = case_when(
    area == "AN" ~ "A\u00f1o Nuevo",
    area == "PL" ~ "Point Lobos",
    area == "BL" ~ "Piedras Blancas",
    area == "PB" ~ "Point Buchon"
  ))

adult_vis_simp$Season <- factor(adult_vis_simp$Season, levels = c("Winter", "Spring" ,"Summer", "Fall"))  
adult_vis_simp$Area <- factor(adult_vis_simp$Area, levels = c("A\u00f1o Nuevo", "Point Lobos" ,"Piedras Blancas", "Point Buchon")) 

adult_pooled <- ggplot()+
  scale_color_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  scale_fill_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  geom_col(position = "dodge2", data = adult_vis_simp, aes(x = lag, y = correlation, 
                                                    color = Season, fill = Season))+
  labs(title = "Adult Blue Rockfish Lag Correlation")+
  scale_y_continuous(limits = c(-0.6, 1.0), 
                     breaks = c(-0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0))+
  theme_bw()+
  theme(panel.grid = element_blank(),
        plot.title = element_text(hjust = 0.5))



```


```{r saving images, eval = F, echo = F}

# 
# ggsave(filename = "2021-07-04_adult_blue_top_corr.png", plot = top_adult_2020_dat,
#        path = "C:/Users/erinj/Documents/Thesis/Figures/Lag/2021", dpi = 1000, width = 6,
#         height = 4)

ggsave(filename = "chpt1_adult_blue_pooled_dat.png", plot = adult_pooled,
       path = "C:/Users/erinj/Documents/Thesis/Figures/chpt1_figs", 
       dpi = 1000, width = 6, height = 4)


# write_csv(top_juv, here("Data", "2021-07-04_top_juv_dataframe.csv"))
# write_csv(top_adult, here("Data", "2021-07-04_top_adult_dataframe.csv"))

```


[back](https://erinmarjo.github.io/CCFRP/27_thesis_progress_report.html)

```{r, eval = F, echo=F}

# Run this in console: rmarkdown::render(input = here("Code", "31_blue_cross_corr_2020.Rmd"), output_dir = here("docs"))

```


```{r}

juv_moci_lags_atmpt <- juv_moci %>%
  filter(area == "AN", site == "MPA", season == "JAS")%>%
  mutate(central_ca1 = dplyr::lag(central_ca, n = 1),
         central_ca2 = dplyr::lag(central_ca, n = 2),
         central_ca3 = dplyr::lag(central_ca, n = 3),
         central_ca4 = dplyr::lag(central_ca, n = 4),
         central_ca5 = dplyr::lag(central_ca, n = 5),
         central_ca6 = dplyr::lag(central_ca, n = 6),
         central_ca7 = dplyr::lag(central_ca, n = 7),
         central_ca8 = dplyr::lag(central_ca, n = 8))


slr_lags <- ggplot(data = juv_moci_lags_atmpt)+
  geom_point(aes(x = cpue_site, y = central_ca), color = "#0000FF")+
  geom_smooth(aes(x = cpue_site, y = central_ca), color = "#0000FF", method = "lm", se = FALSE)+
  geom_point(aes(x = cpue_site, y = central_ca1), color = "#98F5FF")+
  geom_smooth(aes(x = cpue_site, y = central_ca1), color = "#98F5FF", method = "lm", se = FALSE)+
  geom_point(aes(x = cpue_site, y = central_ca2), color = "#00008B")+
  geom_smooth(aes(x = cpue_site, y = central_ca2), color = "#00008B", method = "lm", se = FALSE, size = 2)+
  geom_point(aes(x = cpue_site, y = central_ca3), color = "burlywood1")+
  geom_smooth(aes(x = cpue_site, y = central_ca3), color = "burlywood1", method = "lm", se = FALSE)+
#   geom_point(aes(x = cpue_site, y = central_ca4), color = "cyan")+
#   geom_smooth(aes(x = cpue_site, y = central_ca4), color = "cyan", method = "lm", se = FALSE)+
#    geom_point(aes(x = cpue_site, y = central_ca5), color = "darkgreen")+
#   geom_smooth(aes(x = cpue_site, y = central_ca5), color = "darkgreen", method = "lm", se = FALSE)+
#    geom_point(aes(x = cpue_site, y = central_ca6), color = "orange")+
#   geom_smooth(aes(x = cpue_site, y = central_ca6), color = "orange", method = "lm", se = FALSE)+
#   geom_point(aes(x = cpue_site, y = central_ca7), color = "darkblue")+
#   geom_smooth(aes(x = cpue_site, y = central_ca7), color = "darkblue", method = "lm", se = FALSE)+
#   geom_point(aes(x = cpue_site, y = central_ca8), color = "yellow3")+
#   geom_smooth( aes(x = cpue_site, y = central_ca8), color = "yellow3", method = "lm", se = FALSE)+
  theme_bw()+
  theme(panel.grid = element_blank())

## plot slope. maybe abandon this.

```




Graveyard/Different Analyses

```{r teen loop, echo = F, eval = F}

## Teen Blue Rockfish

# for(j in area_list){
#   for(k in site_list){
#     for(l in season_list){
#       teenager <- teen_moci %>%
#         filter(area == j, site == k, season == l)
#       for(i in 0:8){
#         lag_df <- teenager %>%
#           mutate(central_ca = dplyr::lag(central_ca, n = i)) %>% drop_na()
#         cor <- erin_ccf(teenager$central_ca, teenager$cpue_site,
#                         lag_df$central_ca, lag_df$cpue_site)
#         for(m in 1:length(cor)){
#           teen_corr <- bind_rows(teen_corr, c(area = j, site = k, season = l, 
#                                             lag = -i, correlation = cor[m]))
#         }
#       }
#     }
#   }
# } 
# 
# 
# teen_vis <- teen_corr %>%
#   mutate(correlation = as.numeric(correlation),
#          lag = as.numeric(lag),
#          Season = case_when(
#     season == "JFM" ~ "Winter",
#     season == "AMJ" ~ "Spring",
#     season == "JAS" ~ "Summer",
#     season == "OND" ~ "Fall"
#   ), Area = case_when(
#     area == "AN" ~ "A\u00f1o Nuevo",
#     area == "PL" ~ "Point Lobos",
#     area == "BL" ~ "Piedras Blancas",
#     area == "PB" ~ "Point Buchon"
#   ))
#   
# teen_vis$season <- factor(teen_vis$season, levels = c("JFM", "AMJ" ,"JAS", "OND"))
# teen_vis$Season <- factor(teen_vis$Season, levels = c("Winter", "Spring" ,"Summer", "Fall"))  
# teen_vis$Area <- factor(teen_vis$Area, levels = c("A\u00f1o Nuevo", "Point Lobos" ,"Piedras Blancas", "Point Buchon"))  
# 
# 
# teen_plot_erin_ccf <- ggplot()+
#   scale_color_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
#   scale_fill_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
#   geom_col(position = "dodge2", data = teen_vis, aes(x = lag, y = correlation, 
#                                                     color = Season, fill = Season))+
#   labs(title = "'Teenage' Blue Rockfish Lag Correlation")+
#   scale_y_continuous(limits = c(-0.6, 1.0), 
#                      breaks = c(-0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0))+
#   theme_bw()+
#   theme(panel.grid = element_blank(),
#         plot.title = element_text(hjust = 0.5))+
#   facet_grid(site~Area)

```



```{r top teen figure, eval = F, echo = F}

### Highest Correlation Figure


# top_teen <- teen_vis %>% 
#   arrange(desc(correlation)) %>% 
#   group_by(area, site) %>% slice(1)
# 
# 
# top_teen_meeting <- top_teen %>%
#   mutate(Season = case_when(
#     season == "JFM" ~ "Winter",
#     season == "AMJ" ~ "Spring",
#     season == "JAS" ~ "Summer",
#     season == "OND" ~ "Fall"
#   ), Lag = case_when(
#     lag == "-4" ~ "4",
#     lag == "-3" ~"3",
#     lag == "-2" ~ "2",
#     lag == "-1" ~ "1"
#   ), Area = case_when(
#     area == "AN" ~ "A\u00f1o Nuevo",
#     area == "PL" ~ "Point Lobos",
#     area == "BL" ~ "Piedras Blancas",
#     area == "PB" ~ "Point Buchon"
#   ))
# 
# top_teen_meeting$area <- factor(top_teen_meeting$area, levels = c("AN", "PL", "BL", "PB"))
# top_teen_meeting$Area <- factor(top_teen_meeting$Area, levels = c("A\u00f1o Nuevo", "Point Lobos", "Piedras Blancas", "Point Buchon"))
# top_teen_meeting$Season <- factor(top_teen_meeting$Season, 
#                                  levels = c("Winter", "Spring", "Summer", "Fall"))
# 
# 
# top_teen_2020_dat <- top_teen_meeting %>%
#   ggplot()+
#   geom_point(size = 4, aes(x = Area, y = correlation, color = Season, 
#                            fill = Season, shape = Lag))+
#   scale_color_manual(values = c("#009999","#666666", "#000000"))+
#   scale_fill_manual(values = c("#009999", "#666666", "#000000"))+
#   scale_shape_manual(values = c(15, 19, 17, 6))+
#   labs(title = "'Teenager' Blue Rockfish Top Correlation CPUE-MOCI", 
#        y = "Correlation\n", x = "\nArea (N to S)", caption = "")+
#   scale_y_continuous(limits = c(0.0, 1.0), 
#                      breaks = c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0))+
#   facet_wrap(.~site)+
#   theme_bw()+
#     theme(panel.grid = element_blank(),
#           plot.title = element_text(hjust = 0.5),
#           plot.caption = element_text(size = 8),
#           axis.text.x = element_text(face = "bold",angle = 45, hjust =1))
# 
# 
# top_teen_2020_dat


```
