==> Code/32_pdo_blue_cross_corr.Rmd <==
---
title: "PDO and Blue Rockfish CPUE Cross Correlation"
author: "Erin Johnston"
date: "7/4/2021"
output: 
  prettydoc::html_pretty:
    theme: cayman
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE,
                      warning = F,
                      message = F)

```

```{r libraries}

library(tidyverse)
library(here)
library(knitr)

```


[back](https://erinmarjo.github.io/CCFRP/27_thesis_progress_report.html)


```{r create juv and adult df}

master <- read_csv(here("Data", "2021-06-28_full_master_2020.csv"))

blue_full <- master %>%
  filter(species == "BLU") 

missing_bl_blue <- data.frame(area = c("BL", "BL", "BL", "BL"),
                              site = c("REF", "MPA", "REF", "MPA"),
                              year = c(2007, 2007, 2015, 2015), 
                              cpue_site = c(NA, NA, NA, NA))

juv <- blue_full %>%
  filter(size <= 21) %>%
  select(-size) %>%
  group_by(drift, trip, area, site, month, day, year, gridcell) %>%
  summarise(cpue_sum = sum(cpue))%>%
  group_by(trip, area, site, month, day, year, gridcell) %>%
  summarise(cpue_date = mean(cpue_sum))%>%
  group_by( area, site, year, gridcell) %>%
  summarise(cpue_cell = mean(cpue_date))%>%
  group_by( area, site, year) %>%
  summarise(cpue_site = mean(cpue_cell)) %>%
  bind_rows(missing_bl_blue)


adult <- blue_full %>%
  filter(size >= 32) %>%
  select(-size)%>%
  group_by(drift, trip, area, site, month, day, year, gridcell) %>%
  summarise(cpue_sum = sum(cpue))%>%
  group_by(trip, area, site, month, year, gridcell) %>%
  summarise(cpue_date = mean(cpue_sum)) %>%
  group_by( area, site, year, gridcell) %>%
  summarise(cpue_cell = mean(cpue_date)) %>%
  group_by( area, site, year) %>%
  summarise(cpue_site = mean(cpue_cell))%>%
  bind_rows(missing_bl_blue)

```

## Pacific Decadal Oscillation

This PDO index can be found at this [NOAA website](https://www.ncdc.noaa.gov/teleconnections/pdo/).

```{r pdo data}

pdo <- read_csv(here("Data", "pdo_noaa_ncei_2020.csv")) %>%
  transform( year = substr(date, 1,4), month = substr(date, 5, 6))%>%
  filter(year %in% c(2007:2020))%>%
  mutate(season = case_when(
    month %in% c("01", "02", "03") ~ "JFM",
    month %in% c("04", "05", "06") ~ "AMJ",
    month %in% c("07", "08", "09") ~ "JAS",
    month %in% c("10", "11", "12") ~ "OND")) %>%
  group_by(year, season) %>%
  summarise(pdo_val = mean(value)) %>%
  mutate(year = as.numeric(as.character(year)))


juv_pdo <- left_join(juv, pdo, by = "year")
adult_pdo <- left_join(adult, pdo, by = "year")

```

```{r custom function}

erin_ccf <- function(full_x, full_y, lag_x_rm, lag_y_rm) {
  mean_x <- mean(full_x, na.rm = T)
  mean_y <- mean(full_y, na.rm = T)
  corval <- (sum((lag_x_rm - mean_x)*
                   (lag_y_rm - mean_y)))/
    ((sqrt(sum((full_x - mean_x)^2, na.rm = T)))*
       (sqrt(sum((full_y - mean_y)^2, na.rm = T))))
  return(corval)
  
}

```

```{r loop set up}

juv_corr_pdo <- data.frame(area = character(), site = character(), season = character(), lag = character(), correlation = character())

adult_corr_pdo <- data.frame(area = character(), site = character(), season = character(), lag = character(), correlation = character())

area_list <- c("PB", "BL", "AN", "PL")
site_list <- c("MPA", "REF")
season_list <- c("JFM", "AMJ", "JAS", "OND")

```

## Juvenile Blue Rockfish

```{r juv loop and visual}

for(j in area_list){
  for(k in site_list){
    for(l in season_list){
      juvenile <- juv_pdo %>%
        filter(area == j, site == k, season == l)
      for(i in 0:8){
        lag_df <- juvenile %>%
          mutate(pdo_val = dplyr::lag(pdo_val, n = i)) %>% drop_na()
        cor <- erin_ccf(juvenile$pdo_val, juvenile$cpue_site,
                        lag_df$pdo_val, lag_df$cpue_site)
        for(m in 1:length(cor)){
          juv_corr_pdo <- bind_rows(juv_corr_pdo, c(area = j, site = k, season = l, 
                                            lag = -i, correlation = cor[m]))
        }
      }
    }
  }
} 


juv_vis_pdo <- juv_corr_pdo %>%
  mutate(correlation = as.numeric(correlation),
         lag = as.numeric(lag))
  
juv_vis_pdo$season <- factor(juv_vis_pdo$season, levels = c("JFM", "AMJ" ,"JAS", "OND"))  

juv_plot_pdo <- ggplot()+
  scale_color_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  scale_fill_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  geom_col(position = "dodge2", data = juv_vis_pdo, aes(x = lag, y = correlation, 
                                                    color = season, fill = season))+
  labs(title = "Juvenile Blue Rockfish Lag Correlation: PDO")+
  theme_bw()+
  theme(panel.grid = element_blank(),
        plot.title = element_text(hjust = 0.5))+
  facet_grid(site~area)

juv_plot_pdo

juv_pdo_table <- juv_vis_pdo %>%
  arrange(desc(abs(correlation))) %>%
  group_by(area, site) %>%
  slice(1)

```


```{r juv table}

kable(juv_pdo_table, caption = "Top Juvenile Blue Rockfish Correlation - PDO")

```

```{r juv top plot pdo}

top_juv_pdo <- juv_vis_pdo %>% 
  arrange(desc(abs(correlation))) %>% 
  group_by(area, site) %>% slice(1) %>%
  mutate(lag = as.factor(lag))


# top_juv_pdo <- top_juv_pdo %>%
#   mutate(Season = case_when(
#     season == "JFM" ~ "Winter",
#     season == "AMJ" ~ "Spring",
#     season == "JAS" ~ "Summer",
#     season == "OND" ~ "Fall"
#   ), Lag = case_when(
#     lag == "-2" ~ "2",
#     lag == "-1" ~ "1"
#   ), Area = case_when(
#     area == "AN" ~ "Ano Nuevo",
#     area == "PL" ~ "Point Lobos",
#     area == "BL" ~ "Piedras Blancas",
#     area == "PB" ~ "Point Buchon"
#   ))

top_juv_pdo$area <- factor(top_juv_pdo$area, levels = c("AN", "PL", "BL", "PB"))
# top_juv_pdo$Area <- factor(top_juv_pdo$Area, levels = c("Ano Nuevo", "Point Lobos", "Piedras Blancas", "Point Buchon"))
# top_juv_pdo$Season <- factor(top_juv_pdo$Season, 
#                                  levels = c("Winter", "Spring", "Summer", "Fall"))


top_juv_pdo_plot <- top_juv_pdo %>%
  ggplot()+
  geom_point(size = 4, aes(x = area, y = correlation, color = season, 
                           fill = season, shape = lag))+
   scale_color_manual(values = c("#009999", "#000000"))+
   scale_fill_manual(values = c("#009999", "#000000"))+
  # scale_shape_manual(values = c(15, 19))+
  labs(title = "Juvenile Blue Rockfish Top Correlation CPUE-PDO", 
       y = "Correlation\n", x = "\nArea (N to S)", caption = "")+
  scale_y_continuous(limits = c(0.0, 1.0), 
                     breaks = c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0))+
  facet_wrap(.~site)+
  theme_bw()+
    theme(panel.grid = element_blank(),
          plot.title = element_text(hjust = 0.5),
          plot.caption = element_text(size = 8),
          axis.text.x = element_text(face = "bold",angle = 45, hjust =1))


top_juv_pdo_plot


```



## Adult Blue Rockfish

```{r adult loop and visual}

for(j in area_list){
  for(k in site_list){
    for(l in season_list){
      adult <- adult_pdo %>%
        filter(area == j, site == k, season == l)
      for(i in 0:8){
        lag_df <- adult %>%
          mutate(pdo_val = dplyr::lag(pdo_val, n = i)) %>% drop_na()
        cor <- erin_ccf(adult$pdo_val, adult$cpue_site,
                        lag_df$pdo_val, lag_df$cpue_site)
        for(m in 1:length(cor)){
          adult_corr_pdo <- bind_rows(adult_corr_pdo, c(area = j, site = k, season = l, 
                                            lag = -i, correlation = cor[m]))
        }
      }
    }
  }
} 


adult_vis_pdo <- adult_corr_pdo %>%
  mutate(correlation = as.numeric(correlation),
         lag = as.numeric(lag))
  
adult_vis_pdo$season <- factor(adult_vis_pdo$season, levels = c("JFM", "AMJ" ,"JAS", "OND"))  

adult_plot_pdo <- ggplot()+
  scale_color_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  scale_fill_manual(values = c("#999999", "#009999", "#666666", "#000000"))+
  geom_col(position = "dodge2", data = adult_vis_pdo, aes(x = lag, y = correlation, 
                                                    color = season, fill = season))+
  labs(title = "Adult Blue Rockfish Lag Correlation: PDO")+
  theme_bw()+
  theme(panel.grid = element_blank(),
        plot.title = element_text(hjust = 0.5))+
  facet_grid(site~area)

adult_plot_pdo

adult_pdo_table <- adult_vis_pdo %>%
  arrange(desc(abs(correlation))) %>%
  group_by(area, site) %>%
  slice(1)



```

```{r adult table pdo}

kable(adult_pdo_table, caption = "Top Adult Blue Rockfish Correlation - PDO")

```

```{r adult top plot pdo}

top_adult_pdo <- adult_vis_pdo %>% 
  arrange(desc(abs(correlation))) %>% 
  group_by(area, site) %>% slice(1) %>%
  mutate(lag = as.factor(lag))

top_adult_pdo$area <- factor(top_adult_pdo$area, levels = c("AN", "PL", "BL", "PB"))

top_adult_pdo_plot <- top_adult_pdo %>%
  ggplot()+
  geom_point(size = 4, aes(x = area, y = correlation, color = season, 
                           fill = season, shape = lag))+
    scale_color_manual(values = c("#009999", "#666666"))+
    scale_fill_manual(values = c("#009999", "#666666"))+
  # scale_shape_manual(values = c(15, 19))+
  labs(title = "Adult Blue Rockfish Top Correlation CPUE-PDO", 
       y = "Correlation\n", x = "\nArea (N to S)", caption = "")+
  scale_y_continuous(limits = c(0.0, 1.0), 
                     breaks = c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0))+
  facet_wrap(.~site)+
  theme_bw()+
    theme(panel.grid = element_blank(),
          plot.title = element_text(hjust = 0.5),
          plot.caption = element_text(size = 8),
          axis.text.x = element_text(face = "bold",angle = 45, hjust =1))


top_adult_pdo_plot


```

[back](https://erinmarjo.github.io/CCFRP/27_thesis_progress_report.html)


```{r saving, eval = F, echo = F}

# write_csv(juv_pdo_table, here("Data", "2021-07-05_juv_pdo_topcorr_table.csv"))
# write_csv(adult_pdo_table, here("Data", "2021-07-05_adult_pdo_topcorr_table.csv"))
# 
# 
# ggsave(filename = "2021-07-05_juv_blue_corr_pdo.png", plot = juv_plot_pdo,
#        path = "C:/Users/erinj/Documents/Thesis/Figures/Lag/2021", dpi = 1000, width = 6,
#         height = 4)
# ggsave(filename = "2021-07-05_adult_blue_corr_pdo.png", plot = adult_plot_pdo,
#        path = "C:/Users/erinj/Documents/Thesis/Figures/Lag/2021", dpi = 1000, width = 6,
#         height = 4)
# 
# ggsave(filename = "2021-07-05_juv_blue_top_corr_pdo.png", plot = top_juv_pdo_plot,
#        path = "C:/Users/erinj/Documents/Thesis/Figures/Lag/2021", dpi = 1000, width = 6,
#         height = 4)
# ggsave(filename = "2021-07-05_adult_blue_top_corr_pdo.png", plot = top_adult_pdo_plot,
#        path = "C:/Users/erinj/Documents/Thesis/Figures/Lag/2021", dpi = 1000, width = 6,
#         height = 4)

## run this in console: rmarkdown::render(input = here("Code", "32_pdo_blue_cross_corr.Rmd"), output_dir = here("docs"))

```
