==> Code/33_mhw_growth.Rmd <==
---
title: "MHW Impacts on Adult Growth"
author: "Erin Johnston"
date: "7/6/2021"
output: html_document
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE,
                      message = F,
                      warning = F)

```

```{r}

library(tidyverse)
library(here)
library(ggridges)

```

```{r}

master <- read_csv(here("Data", "2021-06-28_full_master_2020.csv"))

caught_blue <- master %>%
  filter(num_caught > 0, species == "BLU") %>%
  mutate(year = as.factor(year))


hist(caught_blue$size)

caught_adult <- caught_blue %>%
  filter(site == "MPA", area == "BL", size >= 32) %>%
  mutate(size = as.factor(size))


ggplot() +
  geom_histogram(data = caught_adult, aes(x = size), stat = "count")+
  facet_grid(year ~ .) +
  theme_classic()


ggplot() +
  geom_col(data = caught_blue, aes(x = size, y = year, color = site))+
  facet_grid(year~ area)

ggplot()+
  geom_density_ridges(data = caught_blue, aes(x = size, y = year, fill = year), 
                      show.legend = F)+
  geom_vline(xintercept = 21, col = "red")+
  scale_fill_manual(values = c("#003399", "#0000FF", "#3366CC","#006699", "#336666",
                               "#339999","#33CCCC", "#33FFFF", "#00FFCC", 
                               "#00FF66", "#00FF33", "#009933", "#006633", "#666666"))+
  labs(title = "Blue Rockfish Size Density") +
  facet_grid(site~area) +
  theme_bw()+
  theme(panel.grid = element_blank())

```



```{r}
Mode <- function(x) {
  ux <- unique(x)
  ux[which.max(tabulate(match(x, ux)))]
}

getmode <- function(v) {
 uniqv <- unique(v)
 uniqv[which.max(tabulate(match(v, uniqv)))]
}

try <- c(2,3,4,123,123,2,4,5,5,5,5,5,5,5,5,5,6, 6, 6, 6, 6, 6, 6, 6, 10,10,10,10,10,10,10,10,11,11,11,11,11,11,11,11,11, 4,78,9,99,8,77,7,7, 77, 77, 77, 77, 77, 77, 55, 55, 55, 55, 55, 55, 55)



umode <- as.data.frame(try) %>%
  count(try) %>%
  arrange(desc(n)) %>%
  slice_head(n =5) 
umodes2 <- umode$try

nummodes <- 2:5

for(i in nummodes){
  if(abs(umodes2[1] - umodes2[i]) >= 5){
    print(umodes2[i])
  }
  else{
   # print("goodbye")
  }
}

umodes2[1:5]

umodes2[1]- umodes2[3]

print(u$try)

## Find n number of multiple modes
erin_mode <- function(x) {
  u <- as.data.frame(x) %>%
  count(x) %>%
  arrange(desc(n)) %>%
  slice_head(n = 5)
  
  modes <- u$x
  print(modes[1])
  nummodes <- 2:5
  
  for(i in nummodes){
  if(abs(modes[1] - modes[i]) >= 3){
    print(modes[i])
  }
  else{
  }
}
}

erin_mode(try)



umode <- as.data.frame(try) %>%
  count(try) %>%
  arrange(desc(n)) %>%
  slice_head(n =5) %>%
  arrange(try)
umodes2 <- umode$try

nummodes <- 1:4

#umodes2[1+1]

for(i in nummodes){
  if(abs(umodes2[i] - umodes2[i+1]) >= 5){
    print(umodes2[i+1])
  }
  else{
    print("goodbye")
  }
}

for(i in nummodes){
  if(abs(umodes2[1] - umodes2[i+1]) >= 5){
    print(umodes2[i+1])
  }
  else{
    if(abs(umodes2[i]-umodes2[i+1]) >= 5){
      print(umodes2[i+1])
    }
    else{
      print("goodbye")
    }
  }
}

erin_mode2 <- function(x) {
  u <- as.data.frame(x) %>%
    count(x) %>%
    arrange(desc(n)) %>%
    slice_head(n = 5) %>%
    arrange(x)
  
  modes <- u$x
  print(modes[1])
  nummodes <- 1:4
  
  for(i in nummodes){
  if(abs(modes[i] - modes[i+1]) >= 5){
    print(modes[i+1])
  }
  else{
  }
}
}

erin_mode2(try)
```



```{r}

mode_count <- as.data.frame(try) %>%
  count(try) %>%
  arrange(desc(n)) %>%
  slice_head(n = 5)

mod1 <- mode_count %>%
  arrange(try) %>%
  slice_head(n = 1)

mod1try <- mod1$try

mod2 <- mode_count %>%
  filter(try >= mod1try + 5)%>%
  arrange(try) %>%
  slice_head(n = 1)

mod2try <- mod2$try

mod3 <- mode_count %>%
  filter(try >= mod2try + 5) %>%
  arrange(try) %>%
  slice_head(n = 1)

hist(try, breaks = 140)


```



```{r}
Mode(try)

try_blue <- caught_blue %>%
  group_by(year) %>%
  summarise(mode = erin_mode(size))

try_blue0.5 <- caught_blue %>%
  group_by(year) %>%
  summarise(mode = Mode(size))

try_blue2 <- caught_blue %>%
  group_by(year, size) %>%
  summarise(sumsize = sum(num_caught)) %>%
  unite(col = size_num, size:sumsize, sep = "_", remove = F)


try_blue3 <- caught_blue %>%
  group_by(year, size) %>%
  summarise(sumsize = sum(num_caught))

try_blue4 <- caught_blue %>%
  uncount(num_caught)

try_blue5 <- try_blue4 %>%
  group_by(year) %>%
  summarise(mode = Mode(size))




blue_ungroup <- caught_blue %>%
  uncount(num_caught) %>%
  mutate(group = case_when(
    size %in% 5:20 ~ "A",
    size %in% 21:30 ~ "B",
    size %in% 31:40 ~ "C",
    size %in% 41:50 ~ "D"
  ))

caught_blue2 <- caught_blue %>%
  uncount(num_caught)

hist_df <- caught_blue2 %>%
  filter(year == 2015, area == "PB", site == "MPA")


# blue_ungroup <- caught_blue %>%
#   uncount(num_caught) %>%
#   mutate(group = case_when(
#     size %in% 5:9 ~ "A",
#     size %in% 10:14 ~ "B",
#     size %in% 15:19 ~ "C",
#     size %in% 20:24 ~ "D",
#     size %in% 25:29 ~ "E",
#     size %in% 30:34 ~ "F",
#     size %in% 35:39 ~ "G",
#     size %in% 40:44 ~ "H",
#     size %in% 45:49 ~ "I",
#   ))

blue_modes <- blue_ungroup %>%
  group_by(year, group) %>%
  summarise(mode = Mode(size))

caught_blue2 <- caught_blue %>%
  uncount(num_caught) %>%
  mutate(size = as.factor(size))

hist_df <- caught_blue2 %>%
  filter(year == 2015, area == "PB", site == "MPA")

ggplot() +
  geom_histogram(stat = "count", data = hist_df, aes(x = size),
                 col = "grey", binwidth = 1)+
  theme_classic()

hist_df2 <- caught_blue2 %>%
  filter(year %in% c(2015:2020), area == "PB", site == "MPA")

ggplot() +
  geom_histogram(stat = "count", data = hist_df2, aes(x = size),
                 col = "grey", binwidth = 1)+
  facet_grid(year~.) +
  theme_classic()

```


```{r}

library(multimode)

data("geyser")
locmodes(geyser, mod0=2,display=TRUE)

locmodes(try, mod0=2, display = T)

nmodes(try, bw = 5)


blue18 <- caught_blue %>%
  filter(year == 2018, site == "MPA", area == "BL") %>%
  uncount(num_caught)
blue17 <- caught_blue %>%
  filter(year == 2017, site == "MPA", area == "BL") %>%
  uncount(num_caught)


hist(blue17$size, breaks = 45)


hist(blue18$size, breaks = 45)

hist(blue17$size, breaks = 45)
erin_mode2(blue17$size)
erin_mode(blue17$size)

```

```{r}

## just trying my function out with a dataset

datas <- c(1,2,2,3,3,3,3, 3, 3, 3, 3,4,4,4,4,4,4,4,4,5,5,6,6,7,7,8,8,8,8,8,9,9,9,9,9, 10, 11, 11, 11, 12, 12, 13, 13, 13, 13, 14, 14, 14, 14, 15, 16, 16, 17, 17, 17, 17, 17, 17, 17, 17, 17, 18, 18, 18, 18, 18, 18, 18, 19, 19, 19, 19, 20, 20, 20, 21, 21, 22, 22, 22, 22, 22, 22, 22, 22)

hist(datas, breaks =22)

erin_mode(datas)

hist(datas, breaks =22)

erin_mode2(datas)

```



```{r}

library(TropFishR)

data("synLFQ7")

set.seed(1)
synLFQ7a <- lfqModify(synLFQ7, bin_size = 4)


# plot raw and restructured LFQ data
lfqbin <- lfqRestructure(synLFQ7a, MA = 5, addl.sqrt = FALSE)
opar <- par(mfrow = c(2,1), mar = c(2,5,2,3), oma = c(2,0,0,0))
plot(lfqbin, Fname = "catch", date.axis = "modern")
plot(lfqbin, Fname = "rcounts", date.axis = "modern")
par(opar)

```




```{r}

# summarise catch matrix into vector and add plus group which is smaller than Linf
synLFQ7b <- lfqModify(synLFQ7a, vectorise_catch = TRUE, plus_group = 118)
# run catch curve
res_cc <- catchCurve(synLFQ7b, reg_int = c(8,26), calc_ogive = TRUE)
# assign estimates to the data list
synLFQ7b$Z <- res_cc$Z
synLFQ7b$FM <- as.numeric(synLFQ7b$Z - synLFQ7b$M)
synLFQ7b$E <- synLFQ7b$FM/synLFQ7b$Z


```

From TropFishR

```{r}

data(hake)

hake <- hake

VPA(hake, terminalE = 0.5, analysis_type = "VPA", plot = TRUE,
    catch_unit = "'000", plus_group = TRUE)

VPA(hake, terminalE = 0.5, analysis_type = "CA", plot = TRUE,
   catch_unit = "'000", plus_group = TRUE)

## trying to make my own list
midpoints <- c(1,2,3,4,5,6,7,8,9,10)

kval <- 0.5

mydatlist <- list(midpoints, kval)
names(mydatlist) <- c("midpoints", "kvalue")

```
