==> Code/40_mpa_vs_ref_II.Rmd <==
---
title: "New MPA v REF"
author: "Erin Johnston"
date: "9/20/2022"
output: html_document
---

Old file (34_mpa_vs_ref) is causing RStudio to crash upon open. This new file is a copy of some of the MPA v REF work that has been ongoing for the last several months.

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE,
                      warning = F,
                      message = F)

```

```{r packages}

library(tidyverse)
library(here)

```

```{r data}

master <- read_csv(here("Data", "2021-06-28_full_master_2020.csv"))

```


## Update 2022-06-06

We need to add the section on MPA vs REF because it is a piece of the story that is missing. After meeting with Ally, Ben, Josiah on 2022-06-06, I am going to run a glm with the response of cell level cpue for juveniles and then another model with cell level cpue for adults and a separate model for each of the four study areas. I will do this for fixed effect of MOCI negative/positive year and also Before-During-After year (total of 16 models). The fixed effects will be site and some classification of year (see above neg/pos and B/D/A). The models should be the following:

* (juv_cpue ~ site*neg/pos) - 4 models (AN, PL, BL, PB)

* (juv_cpue ~ site*B/D/A) - 4 models (AN, PL, BL, PB)

* (adult_cpue ~ site*neg/pos) - 4 models (AN, PL, BL PB)

* (adult_cpue ~ site*B/D/A) - 4 models (AN, PL, BL PB)

```{r data processing}

blue_full_glm_pro <- master %>%
  filter(species == "BLU") %>%
  mutate(group = as.factor(group),
         trip = as.factor(trip),
         area = as.factor(area),
         month = as.factor(month),
         year = as.factor(year),
         gridcell = as.factor(gridcell),
         site = as.factor(site),
         bda = as.factor(case_when(
           year %in% c(2007:2013) ~ "Before",
           year %in% c(2014:2016) ~ "During",
           year %in% c(2017:2020) ~ "After")),
         moci_state = as.factor(case_when(
           year %in% c(2007:2013) ~ "Negative",
           year %in% c(2014:2020) ~ "Positive"
         )))

# juv <- blue_full_glm_pro %>%
#   filter(size <= 21) %>%
#   select(-size) %>%
#   group_by(drift, trip, area, site, month, day, year, gridcell, bda, moci_state) %>%
#   summarise(cpue_sum = sum(cpue))%>%
#   group_by(trip, area, site, month, day, year, gridcell, bda, moci_state) %>%
#   summarise(cpue_date = mean(cpue_sum))%>%
#   group_by( area, site, year, gridcell, bda, moci_state) %>%
#   summarise(cpue_cell = mean(cpue_date))
# 
# adult <- blue_full_glm_pro %>%
#   filter(size >= 32) %>%
  # select(-size)%>%
  # group_by(drift, trip, area, site, month, day, year, gridcell, bda, moci_state) %>%
  # summarise(cpue_sum = sum(cpue))%>%
  # group_by(trip, area, site, month, year, gridcell, bda, moci_state) %>%
  # summarise(cpue_date = mean(cpue_sum)) %>%
  # group_by( area, site, year, gridcell, bda, moci_state) %>%
  # summarise(cpue_cell = mean(cpue_date))

```


```{r data processing 2, eval = F, echo = F}

# juv_AN <- juv %>%
#   filter(area == "AN")
# 
# juv_PL <- juv %>%
#   filter(area == "PL")
# 
# juv_BL <- juv %>%
#   filter(area == "BL")
# 
# juv_PB <- juv %>%
#   filter(area == "PB")
# 
# adult_AN <- adult %>%
#   filter(area == "AN")
# 
# adult_PL <- adult %>%
#   filter(area == "PL")
# 
# adult_BL <- adult %>%
#   filter(area == "BL")
# 
# adult_PB <- adult %>%
#  filter(area == "PB")

```

Okay, now there are eight datasets, four for adults (one per area) and four for juvenile Blue rockfish (one per area). Now run four sets of four models.


```{r eight juvenile glm models, eval = F, echo = F}

# juv_mod_1 <- glm(cpue_cell ~ site* bda, data = juv_AN)
# summary(juv_mod_1)
# 
# juv_mod_2 <- glm(cpue_cell ~ site*bda, data = juv_PL)
# summary(juv_mod_2)
# 
# juv_mod_3 <- glm(cpue_cell ~ site*bda, data = juv_BL)
# summary(juv_mod_3)
# 
# juv_mod_4 <- glm(cpue_cell ~ site*bda, data = juv_PB)
# summary(juv_mod_4)
# 
# 
# juv_mod_5 <- glm(cpue_cell ~ site*moci_state, data = juv_AN)
# summary(juv_mod_5)
# 
# juv_mod_6 <- glm(cpue_cell ~ site*moci_state, data = juv_PL)
# summary(juv_mod_6)
# 
# juv_mod_7 <- glm(cpue_cell ~ site*moci_state, data = juv_BL)
# summary(juv_mod_7)
# 
# juv_mod_8 <- glm(cpue_cell ~ site*moci_state, data = juv_PB)
# summary(juv_mod_8)

```


```{r eight adult glm models, echo = F, eval = F}

# adult_mod_1 <- glm(cpue_cell ~ site*bda, data = adult_AN)
# summary(adult_mod_1)
# 
# adult_mod_2 <- glm(cpue_cell ~ site*bda, data = adult_PL)
# summary(adult_mod_2)
# 
# adult_mod_3 <- glm(cpue_cell ~ site*bda, data = adult_BL)
# summary(adult_mod_3)
# 
# adult_mod_4 <- glm(cpue_cell ~ site*bda, data = adult_PB)
# summary(adult_mod_4)
# 
# adult_mod_5 <- glm(cpue_cell ~ site*moci_state, data = adult_AN)
# summary(adult_mod_5)
# 
# adult_mod_6 <- glm(cpue_cell ~ site*moci_state, data = adult_PL)
# summary(adult_mod_6)
# 
# adult_mod_7 <- glm(cpue_cell ~ site*moci_state, data = adult_BL)
# summary(adult_mod_7)
# 
# adult_mod_8 <- glm(cpue_cell ~ site*moci_state, data = adult_PB)
# summary(adult_mod_8)


```


## Update 2022-09-20

An email thread ('Statistics are confusing - help!') with Hunter from June 2022 indicates that trying one model with age and area included is the way to go.  Hunter says: 

"So, in my experience and knowledge, the effects of other variables on cpue will not be easier seen if there are separate models for each area and age and they will not be washed out by virtue of the alternative being a single model with area in it. My general recommendation is to use a single model with the entire dataset for at least the following two big reasons:

1. Using a single model gives you the ability to model possible interactions between the variables you're currently subsetting by (area and age), which cannot be done with multiple models. I'm not sure if you have hypotheses about interactions, but this is a big benefit.


2. By using the entire dataset with a single model you will have greater power to detect effects, which is precisely what I think you're trying to do."

I need to check regression assumtions before really moving forward, but I think I need some help with that. My grasp of statistics is slipping the further away from classes I get.


```{r single model data processing}

## make columns that label adult, juvenile, teen. Take out teen for the analysis

single_glm_dat <- blue_full_glm_pro %>%
  mutate(life_stage = as.factor(case_when(
    size <= 21 ~ "Juvenile",
    size >= 32 ~ "Adult",
    size %in% c(22:31) ~ "Teen"))) %>%
  filter(!life_stage == "Teen")


single_glm_dat2 <- single_glm_dat %>%
  select(-size) %>%
  group_by(drift, trip, area, site, month, day, year, gridcell, bda, moci_state, life_stage) %>%
  summarise(cpue_sum = sum(cpue))%>%
  group_by(trip, area, site, month, day, year, gridcell, bda, moci_state, life_stage) %>%
  summarise(cpue_date = mean(cpue_sum))%>%
  group_by( area, site, year, gridcell, bda, moci_state, life_stage) %>%
  summarise(cpue_cell = mean(cpue_date)) %>%
  dplyr::mutate(year = as.numeric(as.character(year))) 



```


# Update 2022-11-14

After meeting with Hunter and receving **

```{r mparef analysis dat}

MR_dat <- single_glm_dat2 %>%
  mutate(year = as.factor(year))

#str(MR_dat)

```

```{r mpa ref analysis model, eval=F}

library(lme4)
library(lmerTest)

MR_mod <- lmerTest::lmer(cpue_cell ~ site + life_stage + year + life_stage:site + year:site + (1|area), data = MR_dat)

## ben was hesitant about adding year:site interaction. So this one is without that second interaction
MR_mod1 <- lmerTest::lmer(cpue_cell ~ site + life_stage + year + life_stage:site + (1|area), data = MR_dat)

summary(MR_mod)
summary(MR_mod1)

## year life stage

```


2022-11-15: Chatted with AP. Life stages separately is biologically relevant. FINAL ANSWER. Will be running the separate models, but NOT doing an ANOVA

```{r two models final answer}

MR_adult <- MR_dat %>%
  filter(life_stage == "Adult")

MR_mod2 <- lmerTest::lmer(cpue_cell ~ site*year + (1|area), data = MR_adult)
summary(MR_mod2)

MR_juv <- MR_dat %>%
  filter(life_stage == "Juvenile")

MR_mod3 <- lmerTest::lmer(cpue_cell ~ site*year + (1|area), data = MR_juv)
summary(MR_mod3)


```


2022-12-31: Bringing this analysis (above MR_mod2 for adults and MR_mod3 for juveniles) into the draft of the thesis. 


```{r, results= 'asis'}

class(MR_mod2) <- "lmerMod"
class(MR_mod3) <- "lmerMod"


library(stargazer)

stargazer(MR_mod2, MR_mod3, type = "html",
          column.labels = c("Adult", "Juvenile"),
          dep.var.labels = "Catch Per Unit Effort (CPUE): Cell Level",
          covariate.labels = c("Reference (REF)", "2008", "2009", "2010", "2011", "2012", "2013", "2014", "2015", "2016", "2017", "2018", "2019", "2020", "REF:2008", "REF:2009", "REF:2010", "REF:2011", "REF:2012", "REF:2013", "REF:2014", "REF:2015", "REF:2016", "REF:2017", "REF:2018", "REF:2019", "REF:2020"),
          model.numbers = FALSE,
          title = "MPA-REF Analysis",
          single.row = TRUE,
          no.space = TRUE,
          align = TRUE,
          column.sep.width = "3pt",
          font.size = "small",
          out = here("Output/MPA_REF_Analysis_test.jpg"))




```


2023-01-17: FINAL FINAL (?) MODEL... maybe back to the one with before/during/after warm water events ('after' also happens to correspond with 10 years of MPA protection in 2017). This will be the exact same models as above EXCEPT instead of using year (categorical with 15 years), we will be using year GROUPINGS of before: 2007-2013, during: 2014-2016, and after: 2017-2020.


```{r two models final answer bda}
options(contrasts = c("contr.treatment","contr.poly"))

MR_adult$bda <- factor(MR_adult$bda, levels = c("Before", "During", "After"))
MR_juv$bda <- factor(MR_juv$bda, levels = c("Before", "During", "After"))


MR_mod4 <- lmerTest::lmer(cpue_cell ~ site*bda + (1|area), data = MR_adult)
summary(MR_mod4)


MR_mod5 <- lmerTest::lmer(cpue_cell ~ site*bda + (1|area), data = MR_juv)
summary(MR_mod5)

```

```{r two bda models, results= 'asis'}

class(MR_mod4) <- "lmerMod"
class(MR_mod5) <- "lmerMod"



library(stargazer)

stargazer(MR_mod4, MR_mod5, type = "html",
          column.labels = c("Adult", "Juvenile"),
          dep.var.labels = "Catch Per Unit Effort (CPUE): Cell Level",
          covariate.labels = c("Reference (REF)", "During", "After", "REF:During", "REF:After"),
          model.numbers = FALSE,
          title = "MPA-REF Analysis",
          single.row = TRUE,
          no.space = TRUE,
          align = TRUE,
          column.sep.width = "3pt",
          font.size = "small",
          out = here("Output/MPA_REF_Analysis_bda.htm"))




```



## investigating the Zeros in the model

```{r}

zero <- MR_dat %>%
  filter(cpue_cell == 0) %>%
  group_by(life_stage, year) %>%
  count() %>%
  unite("stage_year", life_stage:year, sep = "_", remove = F)


tot <- MR_dat %>%
  group_by(life_stage, year) %>%
  count() %>%
  rename(total_cell = n) %>%
  unite("stage_year", life_stage:year, sep = "_", remove = T) 

percent_zero <- left_join(zero, tot, by = "stage_year") %>%
  mutate(perc_zero = n/total_cell*100)

ggplot() +
  geom_col(position = position_dodge(), data = zero, 
           aes(x = year, y = n, fill = life_stage))

ggplot() +
  geom_col(position = position_dodge(), data = percent_zero, 
           aes(x = year, y = perc_zero, fill = life_stage))

```


Everything below is for models run previous to 2022-11-14. See above for updated code
```{r single model lmer, eval = F}
library(lme4)
library(nlme)
library(lmerTest)


single_dat_yearfac <- single_glm_dat2

### do I want to control the order and why does order matter for lmer?
#single_glm_dat2$bda <- factor(single_glm_dat2$bda, levels = c("Before", "During", "After"))

single_glm_dat2a <- MR_dat %>%
  filter(life_stage == "Adult")

single_glm_dat2b <- MR_dat %>%
  filter(life_stage == "Juvenile")


## glm not right if I want to include random effect of area

## year in this model is now a continuous vars
single_mod_1 <- lmer(cpue_cell ~  site + life_stage + year + (1|area) + site:year + life_stage:site + life_stage:year, data = single_glm_dat2)
summary(single_mod_1)

## examine life stage differently? This would mean that we drop life stage and life stage interactions from the model which would make it much more simple. 
## ignore this model for now
# dos_mod_1 <- lmer(cpue_cell ~ site*year + (1|area), data = single_glm_dat2 %>% filter(life_stage == "Adult"))
# summary(dos_mod_1)


##### two models #####


#options(contrasts = c("contr.sum","contr.poly")) ## intersteing SS III where the intercept is the grand mean
#options(contrasts = c("contr.treatment","contr.poly")) ## og, where the intercept is the mean of the first (alphabetical) group

###### THIS IS THE WAY ###### 

## adult

# single_glm_dat2a2 <- single_glm_dat2a %>%
#   mutate(yn = case_when(
#     cpue_cell == 0 ~ "zero",
#     TRUE ~ "not zero"
#   )) %>%
#   group_by(site, year) %>%
#   count(yn)




## zero inflation?
## what to do with years?
## do we include life stage in the model? aka single model
## 

## just run this but not anova 2022-11-15

## year as a factor
## adult
dos_mod_1a <- lmerTest::lmer(cpue_cell ~ site*year (1|area), data = single_glm_dat2a)
anova(dos_mod_1a)
summary(dos_mod_1a)

hist(single_glm_dat2a2$cpue_cell)

hist(asin(sqrt(single_glm_dat2a$cpue_cell)))

hist(log(single_glm_dat2a$cpue_cell+0.5))

## juv
dos_mod_1b <- lmerTest::lmer(cpue_cell ~ site*bda + (1|area), data = single_glm_dat2b)
anova(dos_mod_1b) 
summary(dos_mod_1b)


library(multcomp)
## adult
posthoc_adult <- glht(dos_mod_1a, linfct = mcp(bda = 'Tukey'))
summary(posthoc_adult)
## juv
posthoc_juv <- glht(dos_mod_1b, linfct = mcp(bda = 'Tukey'))
summary(posthoc_juv)


## 


### anova using aov() = same thing, so yay?
##  result is same, why are numbers slightly different?
## adult
dos_mod_1a.2 <- aov(cpue_cell ~ site*bda, data = single_glm_dat2a)
summary(dos_mod_1a.2)
TukeyHSD(dos_mod_1a.2, which = "bda")
## juv
dos_mod_1b.2 <- aov(cpue_cell ~ site*bda, data = single_glm_dat2b)
summary(dos_mod_1b.2)
TukeyHSD(dos_mod_1b.2, which = "bda")

## visuals

#library(ggpubr)
## adult
#ggboxplot(single_glm_dat2a, x = "bda", y = "cpue_cell", color = "site")

aov_adult <- single_glm_dat2a
aov_adult$bda <- factor(aov_adult$bda, levels = c("Before", "During", "After"))


aov_adult_plot <- ggplot(data = aov_adult) +
  geom_boxplot(aes(x = bda, y = cpue_cell, color = site))+
  labs(y = "Catch Per Unit Effort (Fish/Angler Hour)", x = "",
       caption = "Adult Blue rockfish")+
  scale_color_manual("Site", values = c("#990000", "#003366"))+
  scale_y_continuous(limits = c(0,20), breaks = c(0, 5, 10, 15, 20))+
  theme_bw()+
  theme(panel.grid = element_blank())

ggplot(data = aov_adult) +
  geom_boxplot(aes(x = site, y = cpue_cell, color = bda))+
  labs(y = "Catch Per Unit Effort (Fish/Angler Hour)", x = "",
       caption = "Adult Blue rockfish")+
  scale_color_manual("Year Group", values = c("#A2B5CD", "#FF4500", "#9BCD9B"))+
  scale_y_continuous(limits = c(0, 20), breaks = c(0, 5, 10, 15, 20))+
  theme_bw()+
  theme(panel.grid = element_blank())

ggplot(data = aov_adult) +
  geom_boxplot(aes(x = bda, y = cpue_cell))+
  labs(y = "Catch Per Unit Effort (Fish/Angler Hour)", x = "",
       caption = "Adult Blue rockfish")+
  scale_y_continuous(limits = c(0,20), breaks = c(0, 5, 10, 15, 20))+
  theme_bw()+
  theme(panel.grid = element_blank())


aov_juv <- single_glm_dat2b
aov_juv$bda <- factor(aov_juv$bda, levels = c("Before", "During", "After"))


aov_juv_plot <- ggplot(data = aov_juv) +
  geom_boxplot(aes(x = bda, y = cpue_cell, color = site))+
  labs(y = "Catch Per Unit Effort (Fish/Angler Hour)", x = "",
       caption = "Juvenile Blue rockfish")+
  scale_color_manual("Site", values = c("#990000", "#003366"))+
  scale_y_continuous(limits = c(0, 20), breaks = c(0, 5, 10, 15, 20))+
  theme_bw()+
  theme(panel.grid = element_blank())


ggplot(data = aov_juv) +
  geom_boxplot(aes(x = site, y = cpue_cell, color = bda))+
  labs(y = "Catch Per Unit Effort (Fish/Angler Hour)", x = "",
       caption = "Juvenile Blue rockfish")+
  scale_color_manual("Year Group", values = c("#A2B5CD", "#FF4500", "#9BCD9B"))+
  scale_y_continuous(limits = c(0, 20), breaks = c(0, 5, 10, 15, 20))+
  theme_bw()+
  theme(panel.grid = element_blank())




## together
aov_bludat <- single_glm_dat2
aov_bludat$bda <- factor(aov_bludat$bda, levels = c("Before", "During", "After"))

ggplot(data = aov_bludat) +
  geom_boxplot(aes(x = life_stage, y = cpue_cell, color = site))+
  labs(y = "Catch Per Unit Effort (Fish/Angler Hour)", x = "",
       caption = "Adult Blue rockfish")+
  scale_y_continuous(limits = c(0,20), breaks = c(0, 5, 10, 15, 20))+
  theme_bw()+
  theme(panel.grid = element_blank())

##saving the vis
#ggsave(filename = "aov_adult_plot.png", plot = aov_adult_plot, path = here("Images"), dpi = 1000, width = 6, height = 4)

###### testing models without interaction. No interaction has a higher AIC. So model with the interaction above is better 
# dos_mod_1aa <- lmerTest::lmer(cpue_cell ~ site + bda + (1|area), data = single_glm_dat2a)
# anova(dos_mod_1a, dos_mod_1aa, refit = FALSE)
# dos_mod_1bb <- lmerTest::lmer(cpue_cell ~ site + bda + (1|area), data = single_glm_dat2b)
# anova(dos_mod_1b, dos_mod_1bb, refit = FALSE)



## read online if I want the p-value I can use lme instead of lmer?? lmer is useful if I need some major computing skills, but not necessary for this level of analysis. I should get the same coefficients for either function.

# s_mod_2 <- lme(cpue_cell ~ site + life_stage + year + site:year + site:life_stage + life_stage:year, random = ~1|area, data = single_glm_dat2)
# summary(s_mod_2)

## adult and juv and two different

## second model with bda?? reverse stepwise including bda and moci_state
## maybe different model for life stage. life stage is critical enough that we would have to include the interaction between life_stage and everything else.
## area site

## cpue or response ratio as the response?
## ping shelby about MPA REF. 

## 1. does it make more sense for juv and adult to be separate? We expect life stage is going to be affected differently by the different variables that are in the model. 
## 2. what should the unit of replecation be? cpue at the gridcell level? Cell nested in site in area, how to set up. Or should we use response ratio of cpue as response variable?
## 3. 

## Hunter notes: doesnt make sense for interaction between moci and site because oceanography acts on a whole area. no physical bound.


```

10/3/22

After re-reading some notes, I remember that I am supposed to be running two-way repeated measure ANOVAs, where the response is CPUE, the repeated measures are time points before, during, and after the MHW. and the treatment is site type. I think I need to run this as 16 different models though. So back to that thought... should email Hunter again, but I am going to set up the analysis below

Working using this website as a guide: http://agroninfotech.blogspot.com/2020/06/two-way-repeated-measures-analysis-in-r.html

Next maybe I will try this guide: https://www.datanovia.com/en/lessons/repeated-measures-anova-in-r/

```{r eight two way repeated measures ANOVA, eval = F}

## juvenile

jmod_AN <- aov(cpue_cell ~ site*bda + Error(gridcell/(site*bda)), data = juv_AN)
summary(jmod_AN)

jmod_PL <- aov(cpue_cell ~ site*bda + Error(gridcell/(site*bda)), data = juv_PL)
summary(jmod_PL)

jmod_BL <- aov(cpue_cell ~ site*bda + Error(gridcell/(site*bda)), data = juv_BL)
summary(jmod_BL)

jmod_PB <- aov(cpue_cell ~ site*bda + Error(gridcell/(site*bda)), data = juv_PB)
summary(jmod_PB)


## adult

amod_AN <- aov(cpue_cell ~ site*bda + Error(gridcell/(site*bda)), data = adult_AN)
summary(amod_AN)

amod_PL <- aov(cpue_cell ~ site*bda + Error(gridcell/(site*bda)), data = adult_PL)
summary(amod_PL)

amod_BL <- aov(cpue_cell ~ site*bda + Error(gridcell/(site*bda)), data = adult_BL)
summary(amod_BL)

amod_PB <- aov(cpue_cell ~ site*bda + Error(gridcell/(site*bda)), data = adult_PB)
summary(amod_PB)


```

This isn't really going to work because if I need to account for natural variability in the error term, I can't do it by grid cell because each grid cell number is site specific.. aka AN01 as the 'id' will only ever be in the reference area because that is where AN01 occurs.. Better to do two models: one for juve and one for adult and have area be in the error term to account for naturally occurring variability between areas without actually testing between area differences. Right? I will set up the other 7 just in case.

```{r juv and adult two way repeated measures ANOVA, eval = F}

adult_dat_aov <- single_glm_dat2 %>%
  filter(life_stage == "Adult")

juv_dat_aov <- single_glm_dat2 %>%
  filter(life_stage == "Juvenile")


jmod_aov <- aov(cpue_cell ~ site*bda + Error(area/(site*bda)), data = juv_dat_aov)
summary(jmod_aov)

amod_aov <- aov(cpue_cell ~ site*bda + Error(area/(site*bda)), data = adult_dat_aov)
summary(amod_aov)

```


The problem with all of this is that I don't know for certain that subjects are sampled again at the next time point (the repeated measure, whatever that is. Year, bda, moci...). So all of the code is looking for subject ID to go into the error term.


```{r rep measures ANOVA new guide, eval = F}
library(ggpubr)
library(rstatix)
## Data preparationb
## the question: does protection status change cpue for blue rockfish over time for both juveniles and adults

## visualization

juv_AN$bda <- factor(juv_AN$bda, levels = c("Before", "During", "After"))

bxp_jAN <- ggboxplot(juv_AN, x = "bda", y = "cpue_cell", color = "site", palette = "jco")

####checking assumptions
## outliers

juv_AN %>% group_by(site, bda) %>% identify_outliers(cpue_cell) ## there are 9 outliers

## normality assumption

juv_AN %>% group_by(site, bda) %>% shapiro_test(cpue_cell) ## not normally distributed

## qq plots

ggqqplot(juv_AN, "cpue_cell", ggtheme = theme_bw())+
  facet_grid(bda ~ site, labeller = "label_both") ## can more or less assume normality, though some of these don't line up perfectly with the line


## yet another anova, but set up slightly differently

juv_AN <- ungroup(juv_AN)
# juv_AN.aov <- juv_AN %>%
#   group_by(area, site,  bda ) %>%
#   summarise(cpue_site = mean(cpue_cell))

jAN.aov <- anova_test(
  data = juv_AN, dv = cpue_cell, wid = gridcell, within = c(site, bda)
)
##... data are not set up the right way. Don't really get how to do it with my data because I need subject ID AGAIN...

```





Below is a take on Shelby's code on response ratios. UPDATE 9/29: I will not be using response ratio of MPA and REF CPUE as the response, because I need to have site in my analysis to truly talk about site level differences in CPUE


Response Ratios

```{r RR dat processing model and vis SZ, eval = F, echo = F}

single_glm_dat3 <- single_glm_dat %>%
  select(-size) %>%
  group_by(drift, trip, area, site, month, day, year, gridcell, bda, moci_state, life_stage) %>%
  summarise(cpue_sum = sum(cpue))%>%
  group_by(trip, area, site, month, day, year, gridcell, bda, moci_state, life_stage) %>%
  summarise(cpue_date = mean(cpue_sum))%>%
  group_by( area, site, year, gridcell, bda, moci_state, life_stage) %>%
  summarise(cpue_cell = mean(cpue_date)) %>%
  group_by(area, site, year, bda, moci_state, life_stage) %>%
  summarise(cpue_site = mean(cpue_cell))

single_glm_dat_rr <- single_glm_dat3 %>%
  select(area, site, bda, moci_state, life_stage, site, cpue_site) %>%
  pivot_wider(names_from = site, values_from = cpue_site) %>%
  mutate(RR = log(MPA/REF)) %>%
  drop_na()

single_glm_dat_rr$area<-factor(single_glm_dat_rr$area, levels=c("AN", "PL", "BL", "PB"))

single_glm_dat_rr2 <- single_glm_dat_rr %>%
  filter(!REF == 0) ## how do I deal with infinity response ratio??

single_glm_rravg <- single_glm_dat_rr2 %>%
  group_by(year, life_stage) %>%
  summarise(meanRR = mean(RR), seRR = std.error(RR),
            lower = mean(RR) - qt(1-alpha/2, (n() - 1))*sd(RR/sqrt(n())),
            upper = mean(RR) + qt(1-alpha/2, (n() - 1))*sd(RR/sqrt(n())))

## gen 95% CI - look this up in R

ggplot(single_glm_rravg, aes(x=year, y=meanRR))+
  geom_smooth(method='lm', color='black', se=FALSE)+
  ylab('Log(BiomassMPA/BiomassREF)')+
  geom_hline(yintercept = 0, linetype="dashed")+
  geom_errorbar(aes(ymax=upper, ymin=lower),width=0)+
  geom_point(data=single_glm_dat_rr2, aes(x=year, y=RR, fill=area, shape=area), size=3, alpha=0.5)+
  geom_point(shape=21,size=5, alpha=0.8, fill='black')+
  scale_fill_manual(values=c("#ffffff", "#d0d0d0","#858585", "#000000" ))+
  scale_shape_manual(values=c(21,22,23,24))+
  #scale_x_continuous(breaks= c(2007,2009,2011,2013,2013,2015,2017,2019))+
#  ggtitle("Central Coast", "Average RRs")+
  facet_grid(life_stage~.)+
  theme(plot.title = element_text(size = 16, face = "bold"))


```


```{r glm with RR, echo = F}
## should maybe look into GAMM?
## if doing RR - we have single value for each area + year. cant get a RR at the cell level. But will obviously have more data at the cell level.

# single_mod_2 <- glm(RR ~ area + life_stage + year,  data = single_glm_dat_rr2)
# summary(single_mod_2)
```
